Use (now supported) U16 in `ParsedParsecAddr` exposed in the bindings

+++ libparsec/src/addr.rs
@@ -11,20 +11,20 @@ pub enum ParseParsecAddrError {
 pub enum ParsedParsecAddr {
     Server {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
     },
     Organization {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
     },
     OrganizationBootstrap {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -32,7 +32,7 @@ pub enum ParsedParsecAddr {
     },
     WorkspacePath {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -42,7 +42,7 @@ pub enum ParsedParsecAddr {
     },
     InvitationUser {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -50,7 +50,7 @@ pub enum ParsedParsecAddr {
     },
     InvitationDevice {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -58,7 +58,7 @@ pub enum ParsedParsecAddr {
     },
     InvitationShamirRecovery {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -66,7 +66,7 @@ pub enum ParsedParsecAddr {
     },
     PkiEnrollment {
         hostname: String,
-        port: u32,
+        port: u16,
         is_default_port: bool,
         use_ssl: bool,
         organization_id: OrganizationID,
@@ -79,7 +79,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
             ParsecActionAddr::OrganizationBootstrap(addr) => {
                 ParsedParsecAddr::OrganizationBootstrap {
                     hostname: addr.hostname().into(),
-                    port: addr.port() as u32,
+                    port: addr.port(),
                     is_default_port: addr.is_default_port(),
                     use_ssl: addr.use_ssl(),
                     organization_id: addr.organization_id().clone(),
@@ -88,7 +88,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
             }
             ParsecActionAddr::WorkspacePath(addr) => ParsedParsecAddr::WorkspacePath {
                 hostname: addr.hostname().into(),
-                port: addr.port() as u32,
+                port: addr.port(),
                 is_default_port: addr.is_default_port(),
                 use_ssl: addr.use_ssl(),
                 organization_id: addr.organization_id().clone(),
@@ -99,7 +99,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
             ParsecActionAddr::Invitation(addr) => match addr.invitation_type() {
                 InvitationType::User => ParsedParsecAddr::InvitationUser {
                     hostname: addr.hostname().into(),
-                    port: addr.port() as u32,
+                    port: addr.port(),
                     is_default_port: addr.is_default_port(),
                     use_ssl: addr.use_ssl(),
                     organization_id: addr.organization_id().clone(),
@@ -107,7 +107,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
                 },
                 InvitationType::Device => ParsedParsecAddr::InvitationDevice {
                     hostname: addr.hostname().into(),
-                    port: addr.port() as u32,
+                    port: addr.port(),
                     is_default_port: addr.is_default_port(),
                     use_ssl: addr.use_ssl(),
                     organization_id: addr.organization_id().clone(),
@@ -115,7 +115,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
                 },
                 InvitationType::ShamirRecovery => ParsedParsecAddr::InvitationShamirRecovery {
                     hostname: addr.hostname().into(),
-                    port: addr.port() as u32,
+                    port: addr.port(),
                     is_default_port: addr.is_default_port(),
                     use_ssl: addr.use_ssl(),
                     organization_id: addr.organization_id().clone(),
@@ -124,7 +124,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
             },
             ParsecActionAddr::PkiEnrollment(addr) => ParsedParsecAddr::PkiEnrollment {
                 hostname: addr.hostname().into(),
-                port: addr.port() as u32,
+                port: addr.port(),
                 is_default_port: addr.is_default_port(),
                 use_ssl: addr.use_ssl(),
                 organization_id: addr.organization_id().clone(),
@@ -133,7 +133,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
     } else if let Ok(addr) = ParsecOrganizationAddr::from_any(url) {
         Ok(ParsedParsecAddr::Organization {
             hostname: addr.hostname().into(),
-            port: addr.port() as u32,
+            port: addr.port(),
             is_default_port: addr.is_default_port(),
             use_ssl: addr.use_ssl(),
             organization_id: addr.organization_id().clone(),
@@ -141,7 +141,7 @@ pub fn parse_parsec_addr(url: &str) -> Result<ParsedParsecAddr, ParseParsecAddrE
     } else if let Ok(addr) = ParsecAddr::from_any(url) {
         Ok(ParsedParsecAddr::Server {
             hostname: addr.hostname().into(),
-            port: addr.port() as u32,
+            port: addr.port(),
             is_default_port: addr.is_default_port(),
             use_ssl: addr.use_ssl(),
         })
